requirements.txt
Django==4.2.7
djangorestframework==3.14.0
stripe==8.6.0
pytest-django==4.7.0
pytest==7.4.3
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
